TECS-84 refactoring

+++ packages/gatsby/src/bootstrap/index.js
@@ -23,7 +23,7 @@ const preferDefault = require(`./prefer-default`)
 // Add `util.promisify` polyfill for old node versions
 require(`util.promisify/shim`)()
 
-const isNewBuild = true
+const incrementalBuild = true
 
 // Show stack trace on unhandled promises.
 process.on(`unhandledRejection`, (reason, p) => {
@@ -135,6 +135,44 @@ module.exports = async (args: BootstrapArgs) => {
   await apiRunnerNode(`onPreInit`, { parentSpan: activity.span })
   activity.end()
 
+  // During builds, delete html and css files from the public directory as we don't want
+  // deleted pages and styles from previous builds to stick around.
+  if (!incrementalBuild && process.env.NODE_ENV === `production`) {
+    activity = report.activityTimer(
+      `delete html and css files from previous builds`,
+      {
+        parentSpan: bootstrapSpan,
+      }
+    )
+    activity.start()
+    await del([
+      `public/**/*.{html,css}`,
+      `!public/page-data/**/*`,
+      `!public/static`,
+      `!public/static/**/*.{html,css}`,
+    ])
+    activity.end()
+  }
+
+  const cacheDirectory = `${program.directory}/.cache`
+  /*
+   * Copy cache state for comparing data
+   */
+  if (fs.existsSync(`${cacheDirectory}/redux-state.json`)) {
+    let activity = report.activityTimer(
+      `Copied redux-state for comparing later`
+    )
+    activity.start()
+    await fs.copy(
+      `${cacheDirectory}/redux-state.json`,
+      `${program.directory}/temp/redux-state-old.json`,
+      {
+        clobber: true,
+      }
+    )
+    activity.end()
+  }
+
   activity = report.activityTimer(`initialize cache`, {
     parentSpan: bootstrapSpan,
   })
@@ -176,60 +214,21 @@ module.exports = async (args: BootstrapArgs) => {
       data
     `)
   }
-  const cacheDirectory = `${program.directory}/.cache`
-  activity.end()
-  /*
-   * Copy cache state for comparing data
-   */
-  if (fs.existsSync(`${cacheDirectory}/redux-state.json`)) {
-    let activity = report.activityTimer(
-      `Copied redux-state for comparing later`
-    )
-    activity.start()
-    await fs.copy(
-      `${cacheDirectory}/redux-state.json`,
-      `${program.directory}/temp/redux-state-old.json`,
-      {
-        clobber: true,
-      }
-    )
-    activity.end()
-  }
-
-  // During builds, delete html and css files from the public directory as we don't want
-  // deleted pages and styles from previous builds to stick around.
-  if (isNewBuild) {
-    activity = report.activityTimer(
-      `delete html and css files from previous builds`,
-      {
-        parentSpan: bootstrapSpan,
-      }
-    )
-    activity.start()
-    await del([
-      `public/**/*.{html,css,js,map}`,
-      `!public/page-data/**/*`,
-      `!public/static`,
-      `!public/static/**/*.{html,css}`,
-    ])
-    activity.end()
-  }
-
   // Removed to force a new .cache on each build
-  //if (!oldPluginsHash || pluginsHash !== oldPluginsHash) {
-  try {
-    // Attempt to empty dir if remove fails,
-    // like when directory is mount point
-    await fs.remove(cacheDirectory).catch(() => fs.emptyDir(cacheDirectory))
-  } catch (e) {
-    report.error(`Failed to remove .cache files.`, e)
+  if (!oldPluginsHash || pluginsHash !== oldPluginsHash) {
+    try {
+      // Attempt to empty dir if remove fails,
+      // like when directory is mount point
+      await fs.remove(cacheDirectory).catch(() => fs.emptyDir(cacheDirectory))
+    } catch (e) {
+      report.error(`Failed to remove .cache files.`, e)
+    }
+    // Tell reducers to delete their data (the store will already have
+    // been loaded from the file system cache).
+    store.dispatch({
+      type: `DELETE_CACHE`,
+    })
   }
-  // Tell reducers to delete their data (the store will already have
-  // been loaded from the file system cache).
-  store.dispatch({
-    type: `DELETE_CACHE`,
-  })
-  //}
 
   // Update the store with the new plugins hash.
   store.dispatch({
@@ -243,7 +242,7 @@ module.exports = async (args: BootstrapArgs) => {
 
   // Ensure the public/static directory
   await fs.ensureDir(`${program.directory}/public/static`)
-
+  activity.end()
   if (process.env.GATSBY_DB_NODES === `loki`) {
     const loki = require(`../db/loki`)
     // Start the nodes database (in memory loki js with interval disk
@@ -265,8 +264,6 @@ module.exports = async (args: BootstrapArgs) => {
     }
     activity.end()
   }
-
-  // remove all files if new build
 
   activity = report.activityTimer(`copy gatsby files`, {
     parentSpan: bootstrapSpan,
